Refactor (apps/pet-shop): Rename `ps-categories-select.component`
