+++ apps/pet-shop/src/app/integrator/ui/ps-categories/ps-categories-select/categories-select.component.spec.ts
@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing'
-import { PSCategoriesSelectComponent } from './categories-select.component'
+import { PSCategoriesSelectComponent } from './ps-categories-select.component'
 
 // TODO: Explore the tests, refactor, {migrate to `vitest`?}.
